Upgrade to bevy 0.7
Of note: Timer is no longer a component, which was a problem as I
initially gave the bird said component. To fix this, the bird struct
now has a timer field, and a default function that initializes its
timer field to the value I previously used, and thought was suitable.